更新服务注册和日志输出

+++ src/Surging.Core/Surging.Core.CPlatform/Address/AddressModel.cs
@@ -16,7 +16,7 @@ namespace Surging.Core.CPlatform.Address
         public abstract EndPoint CreateEndPoint();
 
         [JsonProperty(DefaultValueHandling = DefaultValueHandling.Ignore)]
-        public decimal ProcessorTime { get; set; }
+        public double ProcessorTime { get; set; }
 
         /// <summary>
         /// 重写后的标识。
